+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.6] - 2022-03-08
 ### Added
 - Arguments "lowercase_only" and "uppercase_only" to the "string" function.
 - "string" function argument "type" will now take "all". It will then return a string containing both numbers and letters.